+++ src/Libraries/SmartStore.Data/Mapping/Catalog/CategoryMap.cs
@@ -10,6 +10,8 @@ namespace SmartStore.Data.Mapping.Catalog
             this.ToTable("Category");
             this.HasKey(c => c.Id);
             this.Property(c => c.Name).IsRequired().HasMaxLength(400);
+			this.Property(c => c.FullName).HasMaxLength(400);
+			this.Property(c => c.BottomDescription).IsMaxLength();
 			this.Property(c => c.Description).IsMaxLength();
             this.Property(c => c.MetaKeywords).HasMaxLength(400);
             this.Property(c => c.MetaTitle).HasMaxLength(400);