+++ src/Presentation/SmartStore.Web/Administration/Controllers/CheckoutAttributeController.cs
@@ -120,18 +120,7 @@ namespace SmartStore.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
                 return AccessDeniedView();
 
-            var checkoutAttributes = _checkoutAttributeService.GetAllCheckoutAttributes();
-            var gridModel = new GridModel<CheckoutAttributeModel>
-            {
-                Data = checkoutAttributes.Select(x => 
-                {
-                    var caModel = x.ToModel();
-                    caModel.AttributeControlTypeName = x.AttributeControlType.GetLocalizedEnum(_localizationService, _workContext);
-                    return caModel;
-                }),
-                Total = checkoutAttributes.Count()
-            };
-            return View(gridModel);
+            return View();
         }
 
         [HttpPost, GridAction(EnableCustomBinding = true)]
@@ -140,7 +129,7 @@ namespace SmartStore.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
                 return AccessDeniedView();
 
-            var checkoutAttributes = _checkoutAttributeService.GetAllCheckoutAttributes();
+            var checkoutAttributes = _checkoutAttributeService.GetAllCheckoutAttributes(true);
             var gridModel = new GridModel<CheckoutAttributeModel>
             {
                 Data = checkoutAttributes.Select(x =>
@@ -164,6 +153,8 @@ namespace SmartStore.Admin.Controllers
                 return AccessDeniedView();
 
             var model = new CheckoutAttributeModel();
+			model.IsActive = true;
+
             //locales
             AddLocales(_languageService, model.Locales);
             PrepareCheckoutAttributeModel(model, null, true);