+++ src/Presentation/SmartStore.Web/Administration/Models/Catalog/ProductListModel.cs
@@ -15,6 +15,7 @@ namespace SmartStore.Admin.Models.Catalog
 			AvailableStores = new List<SelectListItem>();
 			AvailableProductTypes = new List<SelectListItem>();
         }
+
         public GridModel<ProductModel> Products { get; set; }
 
         [SmartResourceDisplayName("Admin.Catalog.Products.List.SearchProductName")]
@@ -24,8 +25,14 @@ namespace SmartStore.Admin.Models.Catalog
         [SmartResourceDisplayName("Admin.Catalog.Products.List.SearchCategory")]
         public int SearchCategoryId { get; set; }
 
+		[SmartResourceDisplayName("Admin.Catalog.Products.List.SearchWithoutCategories")]
+		public bool SearchWithoutCategories { get; set; }
+
         [SmartResourceDisplayName("Admin.Catalog.Products.List.SearchManufacturer")]
         public int SearchManufacturerId { get; set; }
+
+		[SmartResourceDisplayName("Admin.Catalog.Products.List.SearchWithoutManufacturers")]
+		public bool SearchWithoutManufacturers { get; set; }
 
 		[SmartResourceDisplayName("Admin.Common.Store.SearchFor")]
 		public int SearchStoreId { get; set; }